Update ArgentiRotations project file to allow flexible versioning for RotationSolverReborn.Basic package
