+++ ArgentiRotations/ArgentiRotations.csproj
@@ -10,7 +10,7 @@
           </PropertyGroup>
   
           <ItemGroup>
-            <PackageReference Include="RotationSolverReborn.Basic" Version="*-*" />
+            <PackageReference Include="RotationSolverReborn.Basic" Version="*-*"/>
           </ItemGroup>
   
           <ItemGroup>
